Make logging more useful and add in notify func
Logging is now more useful and adds the following fields into the logs:
PID (For Autorip Prelude) and Tag (Which describes which script is
running). This *SHOULD* make it easier to determine what is running
where, and why.

Notify just makes it easier to add in lines to notify me.